add scroll to token asset feature
