+++ packages/client/src/components/SafeSettings/components/SignatureThreshold.js
@@ -12,9 +12,9 @@ const SignatureThreshold = ({ treasury }) => {
     openModal(<EditThreshold treasury={treasury} />);
   };
   return (
-    <div id="signatureThreshold">
+    <div>
       <div className="column p-0 mt-5 is-flex is-full">
-        <h4 className="is-size-5">Signature Threshold</h4>
+        <h2>Signature Threshold</h2>
       </div>
       <div className="column p-5 mt-4 mb-5 is-flex is-full rounded-sm border-light has-shadow">
         <div className="mr-5 is-flex">
