+++ packages/client/src/components/SafeTokens/index.js
@@ -3,8 +3,8 @@ import { useLocation, useHistory } from "react-router-dom";
 import { Web3Context } from "contexts/Web3";
 import { useModalContext } from "contexts";
 import { SendTokens } from "components";
+import { EmptyTableWithCTA } from "library/components";
 import VaultTable from "./VaultTable";
-import EmptyVault from "./EmptyVault";
 import { ASSET_TYPES } from "constants/enums";
 const SafeTokens = () => {
   const location = useLocation();
@@ -29,22 +29,30 @@ const SafeTokens = () => {
     );
   };
   const handleManageTokenVaults = () => {
-    history.push(`/safe/${treasuryAddress}/settings#assets`);
+    history.push(`/safe/${treasuryAddress}/settings#tokenAsset`);
   };
   return (
     <div>
       <div className="is-flex is-justify-content-space-between mt-5">
         <h2 className="flex-1">Tokens</h2>
         <div>
-          <button
-            className="button is-secondary is-small"
-            onClick={handleManageTokenVaults}
-          >
-            Manage Token Vaults
-          </button>
+          {vaults.length > 0 && (
+            <button
+              className="button is-secondary is-small"
+              onClick={handleManageTokenVaults}
+            >
+              Manage Token Vaults
+            </button>
+          )}
         </div>
       </div>
-      {vaults.length === 0 && <EmptyVault />}
+      {vaults.length === 0 && (
+        <EmptyTableWithCTA
+          message="This safe doesn't have any tokens"
+          buttonText="Manage Token Vaults"
+          onButtonClick={handleManageTokenVaults}
+        />
+      )}
       {vaults.length > 0 && (
         <VaultTable vaults={vaults} handleSendToken={handleSendToken} />
       )}
