+++ packages/client/src/library/components/index.js
@@ -1,3 +1,4 @@
 export { default as Dropdown } from "./Dropdown";
 export { default as InputAddress } from "./InputAddress";
 export { default as ProgressBar } from "./ProgressBar";
+export { default as EmptyTableWithCTA } from "./EmptyTableWithCTA";
